Moved host to utils and changed X-Forwarded* headers
