+++ utils/host_test.go
@@ -1,4 +1,4 @@
-package runtime
+package utils
 
 import (
 	"os"
